back_end_application/requirements.txt
Django==3.0
django_phone_field==1.8.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.0
